Move method to product resource
