+++ packages/admin/src/Filament/Resources/ProductResource.php
@@ -24,11 +24,15 @@ use Lunar\Admin\Filament\Resources\ProductResource\Widgets\ProductOptionsWidget;
 use Lunar\Admin\Filament\Widgets\Products\VariantSwitcherTable;
 use Lunar\Admin\Support\Forms\Components\Attributes;
 use Lunar\Admin\Support\Forms\Components\Tags as TagsComponent;
+use Lunar\Admin\Support\Forms\Components\TranslatedText as TranslatedTextInput;
 use Lunar\Admin\Support\RelationManagers\ChannelRelationManager;
 use Lunar\Admin\Support\RelationManagers\MediaRelationManager;
 use Lunar\Admin\Support\RelationManagers\PriceRelationManager;
 use Lunar\Admin\Support\Resources\BaseResource;
 use Lunar\Admin\Support\Tables\Columns\TranslatedTextColumn;
+use Lunar\FieldTypes\Text;
+use Lunar\FieldTypes\TranslatedText;
+use Lunar\Models\Attribute;
 use Lunar\Models\Currency;
 use Lunar\Models\Product;
 use Lunar\Models\ProductVariant;
@@ -168,6 +172,21 @@ class ProductResource extends BaseResource
             'min:1',
             "decimal:0,{$currency->decimal_places}",
         ])->required();
+    }
+
+    public static function getBaseNameFormComponent(): Component
+    {
+        $nameType = Attribute::whereHandle('name')
+            ->whereAttributeType(static::getModel())
+            ->first()?->type ?: TranslatedText::class;
+
+        $component = TranslatedTextInput::make('name');
+
+        if ($nameType == Text::class) {
+            $component = Forms\Components\TextInput::make('name');
+        }
+
+        return $component->label(__('lunarpanel::product.form.name.label'))->required();
     }
 
     protected static function getBrandFormComponent(): Component
